backend/requirements.txt
SQLAlchemy==2.0.31
PyMySQL==1.1.1
cryptography==43.0.1
redis==5.0.6
Pillow==10.4.0
python-dotenv==1.0.1
openai==1.35.14
requests==2.32.3
APScheduler==3.10.4
openpyxl==3.1.5
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
pydantic==2.9.2
pydantic[email]==2.9.2
websockets==13.1
PyJWT==2.9.0
slowapi==0.1.9
pytest==8.3.3
pytest-cov==4.1.0
pytest-json-report==1.5.0